Move API classes to separate package
