+++ NMBSJava/src/api/RouteAPI.java
@@ -1,4 +1,4 @@
-package model;
+package api;
 
 import java.util.ArrayList;
