+++ NMBSJava/src/api/RouteberekeningAPI.java
@@ -1,4 +1,4 @@
-package model;
+package api;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
