+++ NMBSJava/src/api/TimeAPI.java
@@ -1,4 +1,4 @@
-package model;
+package api;
 
 import org.json.JSONObject;
