+++ NMBSJava/src/api/TrainAPI.java
@@ -1,4 +1,4 @@
-package model;
+package api;
 
 import org.json.JSONObject;
